requirements.txt
django
requests_oauthlib
grammarbot
django-oauth-toolkit
ion_oauth
django-guardian
djangorestframework
celery==3.1.25
django-windows-tools
django-celery
nltk
django-ckeditor
pytz
scikit-learn
http.client
urlextract
strsimpy
unidecode
eventlet
whitenoise